+++ Changelog.rst
@@ -27,6 +27,8 @@ New features
 * Added support for ECDSA (FIPS 186-3 and RFC6979).
 * Added `RSA.import_key` as an alias to the deprecated `RSA.importKey`
   (same for the `DSA` module).
+* Added `RsaKey.export_key` as an alias to the deprecated `RsaKey.exportKey`
+  (same for the `DsaKey` class).
 
 Resolved issues
 ---------------